+++ src/main/java/kitchenpos/order/dto/response/OrderResponse.java
@@ -1,10 +1,10 @@
-package kitchenpos.dto.response;
+package kitchenpos.order.dto.response;
 
 import java.time.LocalDateTime;
 import java.util.List;
-import kitchenpos.domain.Order;
-import kitchenpos.domain.OrderLineItem;
-import kitchenpos.domain.OrderStatus;
+import kitchenpos.order.domain.Order;
+import kitchenpos.order.domain.OrderLineItem;
+import kitchenpos.order.domain.OrderStatus;
 
 public class OrderResponse {
     private Long id;
