+++ src/test/java/kitchenpos/order/domain/OrderLineItemTest.java
@@ -1,11 +1,11 @@
-package kitchenpos.domain;
+package kitchenpos.order.domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
 
 import org.junit.jupiter.api.Test;
 
-class OrderLineItemRepositoryTest {
+class OrderLineItemTest {
     @Test
     void 생성() {
         OrderLineItem orderLineItem = new OrderLineItem(1L, null, 1L, 1);
