+++ src/test/java/kitchenpos/order/ui/OrderRestControllerTest.java
@@ -1,4 +1,4 @@
-package kitchenpos.ui;
+package kitchenpos.order.ui;
 
 import static kitchenpos.table.ui.TableRestControllerTest.좌석_생성_요청;
 import static kitchenpos.menu.ui.MenuGroupRestControllerTest.메뉴_그룹_생성_요청;
@@ -12,17 +12,17 @@ import java.util.Arrays;
 import java.util.List;
 import kitchenpos.BaseTest;
 import kitchenpos.menu.domain.MenuGroup;
-import kitchenpos.domain.Order;
-import kitchenpos.domain.OrderLineItem;
-import kitchenpos.domain.OrderStatus;
+import kitchenpos.order.domain.Order;
+import kitchenpos.order.domain.OrderLineItem;
+import kitchenpos.order.domain.OrderStatus;
 import kitchenpos.table.domain.OrderTable;
 import kitchenpos.menu.dto.request.MenuProductRequest;
 import kitchenpos.menu.dto.request.MenuRequest;
-import kitchenpos.dto.request.OrderRequest;
-import kitchenpos.dto.request.OrderStatusRequest;
+import kitchenpos.order.dto.request.OrderRequest;
+import kitchenpos.order.dto.request.OrderStatusRequest;
 import kitchenpos.dto.request.ProductRequest;
 import kitchenpos.menu.dto.response.MenuResponse;
-import kitchenpos.dto.response.OrderResponse;
+import kitchenpos.order.dto.response.OrderResponse;
 import kitchenpos.dto.response.ProductResponse;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
